+++ src/main/java/nextstep/subway/repository/StationRepository.java
@@ -1,8 +1,8 @@
-package nextstep.subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package nextstep.subway.repository;
 
 import java.util.List;
+import nextstep.subway.domain.Station;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
     @Override
